refactor: control channel extraction
